Bugfix to remove NULL character frame delimiter from frame body
Previous versions of `AnyEvent::STOMP::Client` would include the `NULL`
octet of a STOMP frame what was used as delimiter to indicate the end of
the frame. According to the STOMP 1.2 specification [1] the `NULL` octet
is not part of the body:

>  The body is then followed by the NULL octet.

This removes the trailing `NULL` octet from STMOP frames when said octet
was used to delimit the very frame.

[1] https://stomp.github.io/stomp-specification-1.2.html